re-organize components into pages
