--- src/CSS/Contact.css
@@ -1,10 +1,0 @@
-.contactus {
-  text-align: center;
-  max-width: 300px;
-  margin: 20px auto;
-}
-Form {
-  border: 1px solid lightgray;
-  border-radius: 10px;
-  padding: 20px;
-}
