--- src/CSS/Logout.css
@@ -1,3 +1,0 @@
-h1 {
-    padding-top: 50px;
-}
