+++ src/index.css
@@ -20,5 +20,3 @@ code {
   --SiteWhite: #FFFFFF;
   --SiteSageGreen: #98D6D3;
 }
-
-
