+++ src/pages/About/Contact.css
@@ -1,0 +1,5 @@
+.contactus {
+  text-align: center;
+  max-width: 300px;
+  margin: 20px auto;
+}
