+++ src/pages/Dashboard/Dashboard.jsx
@@ -14,7 +14,10 @@ function Dashboard() {
       headers: {
         Accept: "application/json",
         "Content-Type": "application/json"
-      }
+      },
+      body: JSON.stringify(
+        {email: newEmail,
+        password: newPassword})
     })
       .then(res => res.json())
       .then(data => {
@@ -25,6 +28,16 @@ function Dashboard() {
         }
         toast.success("Logout Successful!");
       });
+  }
+
+  function changePassword() {
+    fetch("https://new-mwl-backend.herokuapp.com/users/password/:id", {
+      method: "GET",
+      headers: {
+        Accept: "application/json",
+        "Content-Type": "application/json"
+      }
+    })
   }
 
   return (
@@ -65,7 +78,7 @@ function Dashboard() {
                   Update Password:
                   <input class="input-class" type="text" name="email" />
                 </label>
-                <input className="button-19 m-2" type="submit" value="Submit" />
+                <input className="button-19 m-2" type="submit" value="Submit" onClick={changePassword} />
               </form>
             </div>
             {/* User click buttons, gives option to update password
