+++ src/pages/Logout/Logout.css
@@ -1,0 +1,3 @@
+h1 {
+    padding-top: 100px;
+}
