+++ src/pages/Logout/Logout.js
@@ -1,10 +1,5 @@
 import React from 'react';
-import "../CSS/Logout.css";
-
-// Check for Session Storage
-// If Session storage exists, give user option to logout
-//      Once user clicks logout, redirect to user login page
-// If no session storage exists, redirect user to login page
+import "./Logout.css";
 
 function logout() {
     fetch('https://new-mwl-backend.herokuapp.com/logout', {
