+++ src/pages/Search/MovieCard.css
